requirements.txt
# Complete requirements.txt - Updated with all necessary dependencies

streamlit>=1.28.0
requests>=2.28
sumy>=0.9.0
transformers>=4.30.0
sentence-transformers>=2.2.2
scikit-learn>=1.2.0
numpy>=1.24.0
pandas>=1.5.0
gensim>=4.3.0
plotly>=5.15.0
torch>=2.0.0
google-generativeai>=0.3.0
groq>=0.4.0
rouge-score>=0.1.2
nltk>=3.9
bert-score>=0.3.13
sentencepiece>=0.1.99
sacremoses>=0.0.53
regex>=2024.5.15